refactor: no need to provide two common files under http/ (#3183)
